Edit : Drink Menu Pics + main + mainModals.js + messages -> Product.java, ProductController.java, Order.java Prompt : TextProcessingController.java ModbusService.java => 138.102

+++ src/main/java/taco/rbiz/domain/model/Cart.java
@@ -17,8 +17,8 @@ public class Cart {
         items.add(product);
     }
 
-    public double getTotalPrice() {
-        return items.stream().mapToDouble(Product::getPrice).sum();
+    public int getTotalPrice() {
+        return items.stream().mapToInt(Product::getPrice).sum();
     }
 
     public int getTotalQuantity() {
